feat: [IDP-174]: Account NS Integration. Batch delete (#44106)
* feat: [IDP-174]: Account NS Integration. Batch delete

* feat: [IDP-174]: fix ordering

+++ idp-service/service/src/main/java/io/harness/idp/secret/service/EnvironmentSecretServiceImpl.java
@@ -10,10 +10,15 @@ package io.harness.idp.secret.service;
 import static io.harness.k8s.constants.K8sConstants.BACKSTAGE_SECRET;
 import static io.harness.k8s.constants.K8sConstants.DEFAULT_NAMESPACE;
 
+import static java.lang.String.format;
+
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.beans.DecryptedSecretValue;
 import io.harness.eventsframework.entity_crud.EntityChangeDTO;
+import io.harness.exception.InvalidRequestException;
+import io.harness.idp.namespace.beans.dto.Namespace;
+import io.harness.idp.namespace.service.NamespaceService;
 import io.harness.idp.secret.beans.entity.EnvironmentSecretEntity;
 import io.harness.idp.secret.mappers.EnvironmentSecretMapper;
 import io.harness.idp.secret.repositories.EnvironmentSecretRepository;
@@ -21,6 +26,7 @@ import io.harness.k8s.client.K8sClient;
 import io.harness.secretmanagerclient.services.api.SecretManagerClientService;
 import io.harness.spec.server.idp.v1.model.EnvironmentSecret;
 
+import com.google.common.collect.Streams;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import java.util.ArrayList;
@@ -39,9 +45,11 @@ import lombok.extern.slf4j.Slf4j;
 public class EnvironmentSecretServiceImpl implements EnvironmentSecretService {
   private static final String SUCCEEDED = "succeeded";
   private static final String FAILED = "failed";
+  private static final String IDP_NOT_ENABLED = "IDP has not been set up for account [%s]";
   private EnvironmentSecretRepository environmentSecretRepository;
   private K8sClient k8sClient;
   @Named("PRIVILEGED") private SecretManagerClientService ngSecretService;
+  private NamespaceService namespaceService;
 
   @Override
   public Optional<EnvironmentSecret> findByIdAndAccountIdentifier(String identifier, String accountIdentifier) {
@@ -116,9 +124,25 @@ public class EnvironmentSecretServiceImpl implements EnvironmentSecretService {
   }
 
   @Override
-  public void delete(String secretIdentifier, String harnessAccount) {
+  public void delete(String secretIdentifier, String accountIdentifier) throws Exception {
     EnvironmentSecretEntity environmentSecretEntity = EnvironmentSecretEntity.builder().id(secretIdentifier).build();
+    Optional<EnvironmentSecretEntity> envSecretOpt =
+        environmentSecretRepository.findByAccountIdentifierAndSecretIdentifier(secretIdentifier, accountIdentifier);
+    if (envSecretOpt.isEmpty()) {
+      throw new InvalidRequestException(
+          format("Environment secret [%s] not found in account [%s]", secretIdentifier, accountIdentifier));
+    }
+    k8sClient.removeSecretData(getNamespaceForAccount(accountIdentifier), BACKSTAGE_SECRET,
+        Collections.singletonList(envSecretOpt.get().getName()));
     environmentSecretRepository.delete(environmentSecretEntity);
+  }
+
+  @Override
+  public void deleteMulti(List<String> secretIdentifiers, String accountIdentifier) throws Exception {
+    Iterable<EnvironmentSecretEntity> secrets = environmentSecretRepository.findAllById(secretIdentifiers);
+    List<String> envNames = Streams.stream(secrets).map(EnvironmentSecretEntity::getName).collect(Collectors.toList());
+    k8sClient.removeSecretData(getNamespaceForAccount(accountIdentifier), BACKSTAGE_SECRET, envNames);
+    environmentSecretRepository.deleteAllById(secretIdentifiers);
   }
 
   @Override
@@ -148,6 +172,14 @@ public class EnvironmentSecretServiceImpl implements EnvironmentSecretService {
           ngSecretService.getDecryptedSecretValue(accountIdentifier, null, null, secretIdentifier);
       secretData.put(envName, decryptedValue.getDecryptedValue().getBytes());
     }
-    return k8sClient.updateSecretData(DEFAULT_NAMESPACE, BACKSTAGE_SECRET, secretData, false);
+    return k8sClient.updateSecretData(getNamespaceForAccount(accountIdentifier), BACKSTAGE_SECRET, secretData, false);
+  }
+
+  private String getNamespaceForAccount(String accountIdentifier) {
+    Optional<Namespace> namespaceOpt = namespaceService.getNamespaceForAccountIdentifier(accountIdentifier);
+    if (namespaceOpt.isEmpty()) {
+      throw new InvalidRequestException(format(IDP_NOT_ENABLED, accountIdentifier));
+    }
+    return namespaceOpt.get().getNamespace();
   }
 }
